updated profile and user controllers with jwt authorisation logic

+++ src/controllers/profiles_controller.py
@@ -1,16 +1,20 @@
 from models.Profile import Profile
-# from models.User import User
+from models.User import User
 from main import db
 from schemas.ProfileSchema import profile_schema, profiles_schema
-from flask_jwt_extended import jwt_required
-from flask import Blueprint, request, jsonify
+from flask_jwt_extended import jwt_required, get_jwt_identity
+from flask import Blueprint, request, jsonify, abort
 profiles = Blueprint('profiles', __name__, url_prefix="/profiles")
 
 
 @profiles.route("/", methods=["GET"])
 @jwt_required
 def show_profiles():
-    profiles = Profile.query.all()
+    user_id = get_jwt_identity()
+    user = User.query.get(user_id)
+
+    profiles = Profile.query.filter_by(user_id=user.user_id)
+
     return jsonify(profiles_schema.dump(profiles))
 
 
@@ -18,11 +22,18 @@ def show_profiles():
 @jwt_required
 def create_profile():
     profile_fields = profile_schema.load(request.json)
+    user_id = get_jwt_identity()
+
+    user = User.query.get(user_id)
+
+    if not user:
+        return abort(401, description="Invalid user")
+
     new_profile = Profile()
     new_profile.name = profile_fields["name"]
     new_profile.restrictions = profile_fields["restrictions"]
 
-    db.session.add(new_profile)
+    user.profiles.append(new_profile)
     db.session.commit()
 
     return jsonify(profile_schema.dump(new_profile))
@@ -31,8 +42,19 @@ def create_profile():
 @profiles.route("/<int:id>", methods=["PATCH"])
 @jwt_required
 def update_profile(id):
-    profile = Profile.query.filter_by(profile_id=id)
     profile_fields = profile_schema.load(request.json, partial=True)
+    user_id = get_jwt_identity()
+
+    user = User.query.get(user_id)
+
+    if not user:
+        return abort(401, description="Invalid user")
+
+    profile = Profile.query.filter_by(profile_id=id, user_id=user.user_id)
+
+    if profile.count() != 1:
+        return abort(401, description="Unauthorised to update this profile")
+
     profile.update(profile_fields)
     db.session.commit()
 
@@ -42,7 +64,19 @@ def update_profile(id):
 @profiles.route("/<int:id>", methods=["DELETE"])
 @jwt_required
 def delete_profile(id):
-    profile = Profile.query.get(id)
+    user_id = get_jwt_identity()
+    user = User.query.get(user_id)
+
+    if not user:
+        return abort(401, description="Invalid user")
+
+    profile = Profile.query.filter_by(
+        profile_id=id, user_id=user.user_id).first()
+
+    if not profile:
+        return abort(404, description="Profile not found")
+
     db.session.delete(profile)
     db.session.commit()
+
     return jsonify(profile_schema.dump(profile))
